ppvm_stim::executor

Function sample

pub fn sample<T, I, C, F>(
    program: &ExtendedProgram,
    num_shots: usize,
    make_tableau: F,
) -> Result<Vec<Vec<Option<bool>>>, ExecError>
where
    T: Config,
    <<T as Config>::Storage as BitView>::Store: PrimInt,
    C: SparseVector<Complex<T::Coeff>, I> + Debug,
    T::Coeff: One + Zero + Clone + Num + ToPrimitive + Debug + Mul<f64> + PartialOrd<f64> + PartialOrd + Send + Sync,
    Complex<T::Coeff>: Mul<Output = Complex<T::Coeff>> + From<Complex64> + MulAssign + AddAssign + One + ComplexFloat + Copy,
    I: TableauIndex + Debug + Send + Sync,
    F: Fn(usize) -> GeneralizedTableau<T, I, C> + Sync,

Execute many shots, building a fresh tableau for shot i via make_tableau(i).

When the rayon feature is enabled this dispatches to sample_parallel for batches large enough to amortise thread-scheduling overhead, and to sample_serial otherwise. Without the feature it is always serial. Use sample_serial / sample_parallel directly to force one or the other.